Upgrade to Linux Mint 21.2 Victoria
I did an upgrade to Linux Mint 21.2 Victoria from Linux Mint 21.1 Vera using the Update Manager GUI.  First updating the Update Manager, then a theme, and then selecting Update to Linux Mint 21.2 Victoria on the edit menu.

The page was saved using Firefox "Save Page as... - Web Page, complete" which saves a web page and files used in a subdirectory.  The web page saved is a modified version with references pointing to files in the subdirectory instead of the original copy on the web.

The original URL is: https://ak20.lam1.us/gitweb/?p=etckeeper-ak20.git;a=commitdiff;h=0fd2d11